pyproject.toml
[project]
name = "a2a-registry"
version = "0.1.0"
description = "Central discovery and metadata service for multi-agent systems"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "python-dotenv",
]
